refactor(ui): Move product editing and deleting
In line with the repository editing and deleting, move product edit and
delete functions to "Settings" menu option, which will only be shown
for superusers and product admins. Require manual text confirmation for 
deleting a product.
